优化静态资源处理，使用transferFrom提升性能
将文件读取和输出流的写入操作优化为使用transferFrom方法，减少了中间缓冲区的使用，提高了数据传输的效率。同时，将contentLength字段的类型
从int改为long，以支持更大的文件长度。对HttpStaticResourceHandler和
相关响应类进行了调整，以适应这些变化。

+++ smart-http-server/src/main/java/org/smartboot/http/server/HttpResponse.java
@@ -106,9 +106,9 @@ public interface HttpResponse {
      */
     public Collection<String> getHeaderNames();
 
-    void setContentLength(int contentLength);
+    void setContentLength(long contentLength);
 
-    int getContentLength();
+    long getContentLength();
 
     void setContentType(String contentType);
